+++ src/TextDecoding.ts
@@ -1,5 +1,5 @@
 // polyfill from https://github.com/anonyco/FastestSmallestTextEncoderDecoder/blob/master/EncoderDecoderTogether.src.js
-export const textDecode = (inputArrayOrBuffer: any, options?: any): string => {
+export const textDecode = (inputArrayOrBuffer: any, _options?: any): string => {
   if (typeof TextDecoder !== "undefined") {
     return new TextDecoder("utf-8").decode(inputArrayOrBuffer);
   }